chore: squash database migrations about releases into one (1.2.15)
Changelog: None
Ticket: MEN-6771
