+++ store/mongo/migration_1_2_15.go
@@ -33,8 +33,24 @@ type migration_1_2_15 struct {
 	db     string
 }
 
-// Up intrduces special representation of artifact provides and index them.
 func (m *migration_1_2_15) Up(from migrate.Version) error {
+	err := m.createCollectionReleases()
+	if err == nil {
+		err = m.indexReleaseTags()
+	}
+	if err == nil {
+		err = m.indexReleaseUpdateType()
+	}
+	if err == nil {
+		err = m.indexUpdateTypes()
+	}
+	if err == nil {
+		err = m.indexReleaseArtifactsCount()
+	}
+	return err
+}
+
+func (m *migration_1_2_15) createCollectionReleases() error {
 	ctx := context.Background()
 	c := m.client.Database(m.db).Collection(CollectionImages)
 	cr := m.client.Database(m.db).Collection(CollectionReleases)
@@ -92,7 +108,120 @@ func (m *migration_1_2_15) Up(from migrate.Version) error {
 	if err = cursor.Err(); err != nil {
 		return errors.WithMessage(err, "failed to decode artifact")
 	}
+	return nil
+}
 
+func (m *migration_1_2_15) indexReleaseArtifactsCount() error {
+	ctx := context.Background()
+	idxReleases := m.client.
+		Database(m.db).
+		Collection(CollectionReleases).
+		Indexes()
+
+	_, err := idxReleases.CreateOne(ctx, mongo.IndexModel{
+		Keys: bson.D{
+			{
+				Key:   StorageKeyReleaseArtifactsCount,
+				Value: 1,
+			},
+		},
+		Options: mopts.Index().SetName(IndexNameReleaseArtifactsCount),
+	})
+	if err != nil {
+		return fmt.Errorf("mongo(1.2.15): failed to create index: %w", err)
+	}
+
+	collectionReleases := m.client.
+		Database(m.db).
+		Collection(CollectionReleases)
+	_, err = collectionReleases.UpdateMany(
+		ctx,
+		bson.M{},
+		[]bson.M{
+			{
+				"$set": bson.M{
+					"artifacts_count": bson.M{
+						"$size": "$artifacts",
+					},
+				},
+			},
+		},
+	)
+	if err != nil {
+		return fmt.Errorf("mongo(1.2.15): failed to update adrtifact counts: %w", err)
+	}
+
+	return nil
+}
+
+func (m *migration_1_2_15) indexReleaseTags() error {
+	ctx := context.Background()
+	idxReleases := m.client.
+		Database(m.db).
+		Collection(CollectionReleases).
+		Indexes()
+
+	_, err := idxReleases.CreateOne(ctx, mongo.IndexModel{
+		Keys: bson.D{{
+			Key:   StorageKeyReleaseTags,
+			Value: 1,
+		}, {
+			// Sort by modified date by default when querying by tags
+			Key:   StorageKeyReleaseModified,
+			Value: -1,
+		}},
+		Options: mopts.Index().
+			SetName(IndexNameReleaseTags),
+	})
+	if err != nil {
+		return fmt.Errorf("mongo(1.2.15): failed to create index: %w", err)
+	}
+	return nil
+}
+
+func (m *migration_1_2_15) indexReleaseUpdateType() error {
+	ctx := context.Background()
+	idxReleases := m.client.
+		Database(m.db).
+		Collection(CollectionReleases).
+		Indexes()
+
+	_, err := idxReleases.CreateOne(ctx, mongo.IndexModel{
+		Keys: bson.D{
+			{
+				Key:   StorageKeyReleaseArtifactsUpdateTypes,
+				Value: 1,
+			},
+		},
+		Options: mopts.Index().
+			SetName(IndexNameReleaseUpdateTypes).
+			SetSparse(true),
+	})
+	if err != nil {
+		return fmt.Errorf("mongo(1.2.17): failed to create index: %w", err)
+	}
+	return nil
+}
+
+func (m *migration_1_2_15) indexUpdateTypes() error {
+	ctx := context.Background()
+	idxReleases := m.client.
+		Database(m.db).
+		Collection(CollectionUpdateTypes).
+		Indexes()
+
+	_, err := idxReleases.CreateOne(ctx, mongo.IndexModel{
+		Keys: bson.D{
+			{
+				Key:   StorageKeyTenantId,
+				Value: 1,
+			},
+		},
+		Options: mopts.Index().SetName(IndexNameAggregatedUpdateTypes).SetUnique(true),
+	})
+	if err != nil {
+		return fmt.Errorf("mongo(1.2.18): failed to create index: %w", err)
+	}
 	return nil
 }
