+++ store/mongo/migration_1_2_15_test.go
@@ -26,9 +26,9 @@ import (
 	mstore "github.com/mendersoftware/go-lib-micro/store"
 )
 
-func TestMigration_1_2_15(t *testing.T) {
+func TestMigration_1_2_15_createCollectionReleases(t *testing.T) {
 	if testing.Short() {
-		t.Skip("skipping TestMigration_1_2_15 in short mode.")
+		t.Skip("skipping TestMigration_1_2_15_createCollectionReleases in short mode.")
 	}
 
 	db.Wipe()
@@ -91,6 +91,7 @@ func TestMigration_1_2_15(t *testing.T) {
 					},
 				},
 			},
+			ArtifactsCount: 2,
 		},
 		{
 			Name: "bar",
@@ -105,6 +106,7 @@ func TestMigration_1_2_15(t *testing.T) {
 					},
 				},
 			},
+			ArtifactsCount: 1,
 		},
 	}
 	// insert images
@@ -140,4 +142,241 @@ func TestMigration_1_2_15(t *testing.T) {
 		releases[i].Modified = nil
 	}
 	assert.Equal(t, outputReleases, releases)
+}
+
+func TestMigration_1_2_15_indexReleaseTags(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping TestMigration_1_2_15_indexReleaseTags in short mode.")
+	}
+
+	db.Wipe()
+	c := db.Client()
+
+	ctx := context.TODO()
+
+	//store := NewDataStoreMongoWithClient(c)
+	database := c.Database(mstore.DbFromContext(ctx, DatabaseName))
+	collRel := database.Collection(CollectionReleases)
+
+	// apply migration (1.2.15)
+	mnew := &migration_1_2_15{
+		client: c,
+		db:     DbName,
+	}
+	err := mnew.Up(migrate.MakeVersion(1, 2, 15))
+	assert.NoError(t, err)
+
+	indices := collRel.Indexes()
+	exists, err := hasIndex(ctx, IndexNameReleaseTags, indices)
+	assert.NoError(t, err)
+	assert.True(t, exists, "index "+IndexNameReleaseTags+" must exist in 1.2.15")
+}
+
+func TestMigration_1_2_15_indexReleaseUpdateType(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping TestMigration_1_2_15_indexReleaseUpdateType in short mode.")
+	}
+
+	db.Wipe()
+	c := db.Client()
+
+	ctx := context.TODO()
+
+	//store := NewDataStoreMongoWithClient(c)
+	database := c.Database(mstore.DbFromContext(ctx, DatabaseName))
+	collRel := database.Collection(CollectionReleases)
+
+	// apply migration (1.2.15)
+	mnew := &migration_1_2_15{
+		client: c,
+		db:     DbName,
+	}
+	err := mnew.Up(migrate.MakeVersion(1, 2, 15))
+	assert.NoError(t, err)
+
+	indices := collRel.Indexes()
+	exists, err := hasIndex(ctx, IndexNameReleaseUpdateTypes, indices)
+	assert.NoError(t, err)
+	assert.True(t, exists, "index "+IndexNameReleaseUpdateTypes+" must exist in 1.2.15")
+}
+
+func TestMigration_1_2_15_indexUpdateTypes(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping TestMigration_1_2_15_indexUpdateTypes in short mode.")
+	}
+
+	db.Wipe()
+	c := db.Client()
+
+	ctx := context.TODO()
+
+	//store := NewDataStoreMongoWithClient(c)
+	database := c.Database(mstore.DbFromContext(ctx, DatabaseName))
+	collRel := database.Collection(CollectionUpdateTypes)
+
+	// apply migration (1.2.15)
+	mnew := &migration_1_2_15{
+		client: c,
+		db:     DbName,
+	}
+	err := mnew.Up(migrate.MakeVersion(1, 2, 18))
+	assert.NoError(t, err)
+
+	indices := collRel.Indexes()
+	exists, err := hasIndex(ctx, IndexNameAggregatedUpdateTypes, indices)
+	assert.NoError(t, err)
+	assert.True(t, exists, "index "+IndexNameAggregatedUpdateTypes+" must exist in 1.2.18")
+}
+
+func TestMigration_1_2_15_indexReleaseArtifactsCount(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping TestMigration_1_2_15_indexReleaseArtifactsCount in short mode.")
+	}
+
+	db.Wipe()
+	c := db.Client()
+
+	ctx := context.TODO()
+
+	//store := NewDataStoreMongoWithClient(c)
+	database := c.Database(mstore.DbFromContext(ctx, DatabaseName))
+	collRel := database.Collection(CollectionReleases)
+
+	artifactType := "app"
+	inputImgs := []*model.Image{
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d80",
+			ImageMeta: &model.ImageMeta{
+				Description: "description",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App1 v1.0",
+				DeviceTypesCompatible: []string{"foo"},
+			},
+			Modified: timePtr("2010-09-22T22:00:00+00:00"),
+		},
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d81",
+			ImageMeta: &model.ImageMeta{
+				Description: "description",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App2 v0.1",
+				DeviceTypesCompatible: []string{"foo"},
+				Updates:               []model.Update{},
+			},
+			Modified: timePtr("2010-09-22T23:02:00+00:00"),
+		},
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d82",
+			ImageMeta: &model.ImageMeta{
+				Description: "description",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App1 v1.0",
+				DeviceTypesCompatible: []string{"bar, baz"},
+				Updates:               []model.Update{},
+			},
+			Modified: timePtr("2010-09-22T22:00:01+00:00"),
+		},
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d83",
+			ImageMeta: &model.ImageMeta{
+				Description: "description",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App1 v1.0",
+				DeviceTypesCompatible: []string{"bork"},
+				Updates:               []model.Update{},
+			},
+			Modified: timePtr("2010-09-22T22:00:04+00:00"),
+		},
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d84",
+			ImageMeta: &model.ImageMeta{
+				Description: "extended description",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App2 v0.1",
+				DeviceTypesCompatible: []string{"bar", "baz"},
+				Updates:               []model.Update{},
+			},
+			Modified: timePtr("2010-09-22T23:00:00+00:00"),
+		},
+		{
+			Id: "6d4f6e27-c3bb-438c-ad9c-d9de30e59d85",
+			ImageMeta: &model.ImageMeta{
+				Description: "description2",
+			},
+
+			ArtifactMeta: &model.ArtifactMeta{
+				Name:                  "App4 v2.0",
+				DeviceTypesCompatible: []string{"foo2"},
+				Updates: []model.Update{
+					{
+						TypeInfo: model.ArtifactUpdateTypeInfo{
+							Type: &artifactType,
+						},
+					},
+				},
+			},
+			Modified: timePtr("2023-09-22T22:00:00+00:00"),
+		},
+	}
+	releases := []model.Release{
+		{
+			Name: "App1 v1.0",
+			Artifacts: []model.Image{
+				*inputImgs[0],
+				*inputImgs[2],
+				*inputImgs[3],
+			},
+		},
+		{
+			Name: "App2 v0.1",
+			Artifacts: []model.Image{
+				*inputImgs[1],
+				*inputImgs[4],
+			},
+		},
+		{
+			Name: "App4 v2.0",
+			Artifacts: []model.Image{
+				*inputImgs[5],
+			},
+		},
+	}
+	items := make([]interface{}, len(releases))
+	for i, _ := range releases {
+		items[i] = releases[i]
+	}
+	r, err := collRel.InsertMany(ctx, items)
+	assert.NotNil(t, r)
+	assert.NoError(t, err)
+
+	// apply migration (1.2.15)
+	mnew := &migration_1_2_15{
+		client: c,
+		db:     DbName,
+	}
+	err = mnew.Up(migrate.MakeVersion(1, 2, 15))
+	assert.NoError(t, err)
+
+	indices := collRel.Indexes()
+	exists, err := hasIndex(ctx, IndexNameReleaseArtifactsCount, indices)
+	assert.NoError(t, err)
+	assert.True(t, exists, "index "+IndexNameAggregatedUpdateTypes+" must exist in 1.2.15")
+
+	cursor, err := collRel.Find(ctx, bson.M{})
+	var releases010219 []model.Release
+	err = cursor.All(ctx, &releases010219)
+	assert.NoError(t, err)
+	for _, r := range releases010219 {
+		assert.Equal(t, len(r.Artifacts), r.ArtifactsCount)
+	}
 }
